refactor(Lotto) : 상속 이용하여 분리
- LottoTicket
- WinLottoTicket
