+++ src/main/java/lotto/domain/item/WinLottoTicket.java
@@ -1,0 +1,4 @@
+package lotto.domain.item;
+
+public class WinLottoTicket {
+}
